Moved config.json into config directory.
