+++ src/cmds/vcs_cmds/hooks/webhooks.js
@@ -1,7 +1,7 @@
 /* eslint-disable max-len */
 const request = require('request');
 const authorisify = require('../../../lib/authorisify');
-const config = require('../../../config.json');
+const config = require('../../../config/config.json');
 
 const LINC_API_SITES_ENDPOINT = `${config.Api.LincBaseEndpoint}/sites`;
