+++ src/cmds/webhook_cmds/webhooks.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-param-reassign */
 const request = require('request');
 const authorisify = require('../../lib/authorisify');
-const config = require('../../config.json');
+const config = require('../../config/config.json');
 
 const LINC_API_SITES_ENDPOINT = `${config.Api.LincBaseEndpoint}/sites`;
